Fix code after feedback
