+++ package.json
@@ -16,7 +16,6 @@
     "mongodb": "^5.7.0",
     "mongoose": "^7.4.0",
     "next": "13.2.3",
-    "node-fetch": "^3.3.2",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-icons": "^4.10.1",
